refactor: extract date utilities to shared lib/utils/datetime
