+++ ui/src/hooks/url-state/useMetricsUrlState.ts
@@ -2,16 +2,9 @@ import { useCallback, useState, useEffect, useMemo } from "react";
 
 import { useQueryState, parseAsString } from "nuqs";
 
-import { type SelectionMode, type MetricType } from "./types";
+import { dateToDateTimeLocal } from "@/lib/utils/datetime";
 
-function toLocalDateTimeString(date: Date): string {
-  const y = date.getFullYear();
-  const m = String(date.getMonth() + 1).padStart(2, "0");
-  const d = String(date.getDate()).padStart(2, "0");
-  const h = String(date.getHours()).padStart(2, "0");
-  const min = String(date.getMinutes()).padStart(2, "0");
-  return `${y}-${m}-${d}T${h}:${min}`;
-}
+import { type SelectionMode, type MetricType } from "./types";
 
 interface UseMetricsUrlStateResult {
   selectedChip: string;
@@ -57,8 +50,8 @@ export function useMetricsUrlState(): UseMetricsUrlStateResult {
     const now = new Date();
     const sevenDaysAgo = new Date(now.getTime() - 7 * 24 * 60 * 60 * 1000);
     return {
-      start: toLocalDateTimeString(sevenDaysAgo),
-      end: toLocalDateTimeString(now),
+      start: dateToDateTimeLocal(sevenDaysAgo),
+      end: dateToDateTimeLocal(now),
     };
   }, []);
 
@@ -112,8 +105,8 @@ export function useMetricsUrlState(): UseMetricsUrlStateResult {
     (days: number) => {
       const now = new Date();
       const past = new Date(now.getTime() - days * 24 * 60 * 60 * 1000);
-      setStartDateState(toLocalDateTimeString(past));
-      setEndDateState(toLocalDateTimeString(now));
+      setStartDateState(dateToDateTimeLocal(past));
+      setEndDateState(dateToDateTimeLocal(now));
     },
     [setStartDateState, setEndDateState],
   );
