+++ ui/src/hooks/url-state/useRangeModeUrlState.ts
@@ -2,14 +2,7 @@ import { useCallback, useMemo } from "react";
 
 import { useQueryState, parseAsString } from "nuqs";
 
-function toLocalDateTimeString(date: Date): string {
-  const y = date.getFullYear();
-  const m = String(date.getMonth() + 1).padStart(2, "0");
-  const d = String(date.getDate()).padStart(2, "0");
-  const h = String(date.getHours()).padStart(2, "0");
-  const min = String(date.getMinutes()).padStart(2, "0");
-  return `${y}-${m}-${d}T${h}:${min}`;
-}
+import { dateToDateTimeLocal } from "@/lib/utils/datetime";
 
 interface UseRangeModeUrlStateResult {
   startDate: string;
@@ -27,8 +20,8 @@ export function useRangeModeUrlState(): UseRangeModeUrlStateResult {
     const now = new Date();
     const sevenDaysAgo = new Date(now.getTime() - 7 * 24 * 60 * 60 * 1000);
     return {
-      start: toLocalDateTimeString(sevenDaysAgo),
-      end: toLocalDateTimeString(now),
+      start: dateToDateTimeLocal(sevenDaysAgo),
+      end: dateToDateTimeLocal(now),
     };
   }, []);
 
@@ -50,8 +43,8 @@ export function useRangeModeUrlState(): UseRangeModeUrlStateResult {
     (days: number) => {
       const now = new Date();
       const past = new Date(now.getTime() - days * 24 * 60 * 60 * 1000);
-      setStartDateState(toLocalDateTimeString(past));
-      setEndDateState(toLocalDateTimeString(now));
+      setStartDateState(dateToDateTimeLocal(past));
+      setEndDateState(dateToDateTimeLocal(now));
     },
     [setStartDateState, setEndDateState],
   );
